RDAIngest

Documentation for RDAIngest.

Functions

RDAIngest.AbstractSource

RDAIngest.CHAMPSSource

RDAIngest.COMSASource

RDAIngest.Ingest

RDAIngest.add_data_column

RDAIngest.add_data_column

RDAIngest.add_datarows

RDAIngest.add_domain

RDAIngest.add_ethics

RDAIngest.add_instruments

RDAIngest.add_protocols

RDAIngest.add_sites

RDAIngest.add_sites

RDAIngest.add_source

RDAIngest.add_transformation

RDAIngest.add_variables

RDAIngest.add_variables

RDAIngest.add_vocabulary

RDAIngest.createdatabase

RDAIngest.createdatabasesqlite

RDAIngest.createdatabasesqlserver

RDAIngest.createdatasets

RDAIngest.createdatasets

RDAIngest.createdeaths

RDAIngest.createdeaths

RDAIngest.createinstruments

RDAIngest.createinstruments

RDAIngest.createmapping

RDAIngest.createmapping

RDAIngest.createprotocols

RDAIngest.createprotocols

RDAIngest.createsources

RDAIngest.createsources

RDAIngest.createtransformations

RDAIngest.createtransformations

RDAIngest.createvariables

RDAIngest.createvariables

RDAIngest.dataset_column

RDAIngest.dataset_column

RDAIngest.dataset_in_ingest

RDAIngest.dataset_to_arrow

RDAIngest.dataset_to_csv

RDAIngest.dataset_to_dataframe

RDAIngest.dataset_to_dataframe

RDAIngest.dataset_variables

RDAIngest.dataset_variables

RDAIngest.death_in_ingest

RDAIngest.get_datasetname

RDAIngest.get_datasetname

RDAIngest.get_domain

RDAIngest.get_last_deathingest

RDAIngest.get_last_deathingest

RDAIngest.get_namedkey

RDAIngest.get_source

RDAIngest.get_table

RDAIngest.get_table

RDAIngest.get_valuetype

RDAIngest.get_variable

RDAIngest.get_variable

RDAIngest.get_vocabulary

RDAIngest.ingest_data

RDAIngest.ingest_deaths

RDAIngest.ingest_dictionary

RDAIngest.ingest_source

RDAIngest.ingest_tac_vocabulary

RDAIngest.initstatuses

RDAIngest.inittypes

RDAIngest.initvalue_types

RDAIngest.insertdata

RDAIngest.insertwithidentity

RDAIngest.insertwithidentity

RDAIngest.lines

RDAIngest.link_deathrows

RDAIngest.link_deathrows

RDAIngest.link_instruments

RDAIngest.lookup_variables

RDAIngest.makeodbcparam

RDAIngest.makeparam

RDAIngest.opendatabase

RDAIngest.opensqlitedatabase

RDAIngest.opensqlserverdatabase

RDAIngest.prepareinsertstatement

RDAIngest.prepareinsertstatement

RDAIngest.prepareselectstatement

RDAIngest.prepareselectstatement

RDAIngest.read_data

RDAIngest.read_sitedata

RDAIngest.read_variables

RDAIngest.save_dataset

RDAIngest.savedataframe

RDAIngest.savedataframe

RDAIngest.selectdataframe

RDAIngest.selectdataframe

RDAIngest.selectsourcesites

RDAIngest.selectsourcesites

RDAIngest.updatevariable_vocabulary

RDAIngest.AbstractSource — Type

Provide a source struct for source specific information

RDAIngest.CHAMPSSource — Type

Provide CHAMPS specific information

RDAIngest.COMSASource — Type

Provide CHAMPS specific information

RDAIngest.Ingest — Type

Provide ingest specific information

RDAIngest.add_data_column — Method

add_data_column(db::ODBC.Connection, variable_id, value_type, coldata)

Insert data for a column of the source dataset

RDAIngest.add_data_column — Method

add_data_column(db::SQLite.DB, variable_id, coldata)

Insert data for a column of the source dataset

RDAIngest.add_datarows — Method

add_datarows(db::SQLite.DB, nrow::Integer, dataset_id::Integer)

Define data rows in the datarows table

RDAIngest.add_domain — Function

add_domain(db::DBInterface.Connection, domain_name::String, domain_description::String)

Add domain to the domain table if not exist, and returns the domain id

RDAIngest.add_ethics — Method

add_ethics(source::AbstractSource, db::DBInterface.Connection, datapath::String)

Ethics document, committee and reference need to be in matching order

RDAIngest.add_instruments — Method

add_instruments(source::AbstractSource, db::SQLite.DB, datapath::String)

Add survey instruments

RDAIngest.add_protocols — Method

add_protocols(source::AbstractSource, db::SQLite.DB, datapath::String)

Add protocols Todo: how protocols link to enthics_id, need a mapping dictionary?

RDAIngest.add_sites — Method

add_sites(source::CHAMPSSource, db::DBInterface.Connection, sourceid::Integer, datapath::String)

Add CHAMPS sites and country iso2 codes to sites table

RDAIngest.add_sites — Method

add_sites(source::COMSASource, db::DBInterface.Connection, sourceid::Integer, datapath::String)

Add COMSA sites and country iso2 codes to sites table

RDAIngest.add_source — Method

add_source(source::AbstractSource, db::DBInterface.Connection)

Add source name to the sources table, and returns the source_id

RDAIngest.add_transformation — Method

add_transformation(db::SQLite.DB, type::Integer, status::Integer, description::String, code_reference::String, date_created::Date, created_by::String)

Add a transformation to the transformations table

RDAIngest.add_variables — Method

add_variables(variables::AbstractDataFrame, db::ODBC.Connection, domain_id::Integer)

Add variables from a variable dataframe to variables table

RDAIngest.add_variables — Method

add_variables(variables::AbstractDataFrame, db::SQLite.DB, domain_id::Integer)

Add variables from a variable dataframe to variables table

RDAIngest.add_vocabulary — Method

add_vocabulary(db::SQLite.DB, vocabulary::Vocabulary)

Insert a vocabulary and its items into a RDA database, returns the vocabulary_id of the inserted vocabulary

RDAIngest.createdatabase — Method

createdatabase(path, name; replace=false, sqlite=true)

Creates a database to store the information contained in the Reference Death Archive (RDA) By default a sqlite database is created, but this can be changed by setting the sqlite argument to false, in which case a sql server database is created and path is interpreted as the name of the database server.

RDAIngest.createdatabasesqlite — Method

createdatabasesqlite(path, name; replace=replace)::SQLite.DB

Create an sqlite database on path with name, if replace = true then replace any existing database

RDAIngest.createdatabasesqlserver — Method

createdatabasesqlserver(server, name; replace=replace)::ODBC.Connection

Create a SQL Server database on server with name, if replace = true then replace any existing database

RDAIngest.createdatasets — Method

createdatasets(db::ODBC.Connection)

Create tables to record datasets, rows, data and links to the transformations that use/created the datasets

RDAIngest.createdatasets — Method

createdatasets(db::SQLite.DB)

Create tables to record datasets, rows, data and links to the transformations that use/created the datasets

RDAIngest.createdeaths — Method

createdeaths(db::ODBC.Connection)

Create tables to store deaths, and their association with data rows and data ingests

RDAIngest.createdeaths — Method

createdeaths(db)

Create tables to store deaths, and their association with data rows and data ingests

RDAIngest.createinstruments — Method

createinstruments(db::ODBC.Connection)

Create tables to record data collection instruments, and their associated protocols and datasets

RDAIngest.createinstruments — Method

createinstruments(db::SQLite.DB)

Create tables to record data collection instruments, and their associated protocols and datasets

RDAIngest.createmapping — Method

createmapping(db::ODBC.Connection)

Create the table required for variable mapping. This table is used to map variables from one instrument to another. The table is created in the database provided as an argument. The variable mapping is based on the PyCrossVA approach.

The relationship to the PyCrossVA configuration file columns:

New Column Name = destinationid - the variableid of the new column

New Column Documentation = Stored in the variable table

Source Column ID = fromid - the variableid of the source variable

Source Column Documentation = will be in the variables table

Relationship = operator - the operator to be used to create the new variable

Condition = operants - the operants to be used with the operator

Prerequisite = prerequisiteid - the variableid of the prerequisite variable

RDAIngest.createmapping — Method

createmapping(db::SQLite.DB)

Create the table required for variable mapping. This table is used to map variables from one instrument to another. The table is created in the database provided as an argument. The variable mapping is based on the PyCrossVA approach.

The relationship to the PyCrossVA configuration file columns:

New Column Name = destinationid - the variableid of the new column

New Column Documentation = Stored in the variable table

Source Column ID = fromid - the variableid of the source variable

Source Column Documentation = will be in the variables table

Relationship = operator - the operator to be used to create the new variable

Condition = operants - the operants to be used with the operator

Prerequisite = prerequisiteid - the variableid of the prerequisite variable

RDAIngest.createprotocols — Method

createprotocols(db::ODBC.Connection)

Create tables to record information about protocols and the ethics approvals for those protocols

RDAIngest.createprotocols — Method

createprotocols(db::SQLite.DB)

Create tables to record information about protocols and the ethics approvals for those protocols

RDAIngest.createsources — Method

createsources(db::ODBC.Connection)

Creates tables to record a source and associated site/s for deaths contributed to the RDA

RDAIngest.createsources — Method

createsources(db::SQLite.DB)

Creates tables to record a source and associated site/s for deaths contributed to the RDA

RDAIngest.createtransformations — Method

createtransformations(db::ODBC.Connection)

Create tables to record data transformations and data ingests

RDAIngest.createtransformations — Method

createtransformations(db)

Create tables to record data transformations and data ingests

RDAIngest.createvariables — Method

createvariables(db::ODBC.Connection)

Create tables to record value types, variables and vocabularies

RDAIngest.createvariables — Method

createvariables(db)

Create tables to record value types, variables and vocabularies

RDAIngest.dataset_column — Method

dataset_column(db::ODBC.Connection, dataset_id::Integer, variable_id::Integer, variable_name::String)::AbstractDataFrame

Return one column of data in a dataset (representing a variable)

RDAIngest.dataset_column — Method

dataset_column(db::SQLite.DB, dataset_id::Integer, variable_id::Integer, variable_name::String)::AbstractDataFrame

Return one column of data in a dataset (representing a variable)

RDAIngest.dataset_in_ingest — Method

datasetiningest(db, datasetid, ingestionid)

If datasetid is part of ingestionid

RDAIngest.dataset_to_arrow — Method

dataset_to_arrow(db, dataset, datapath)

Save a dataset in the arrow format

RDAIngest.dataset_to_csv — Function

dataset_to_csv(db, dataset_id, datapath, compress)

Save a dataset in compressed csv format

RDAIngest.dataset_to_dataframe — Method

dataset_to_dataframe(db::ODBC.Connection, dataset)::AbstractDataFrame

Extract a dataset into a DataFrame

RDAIngest.dataset_to_dataframe — Method

dataset_to_dataframe(db::SQLite.DB, dataset)::AbstractDataFrame

Return a dataset with id dataset as a DataFrame in the wide format

RDAIngest.dataset_variables — Method

dataset_variables(db::ODBC.Connection, dataset)::AbstractDataFrame

Return the list of variables in a dataset

RDAIngest.dataset_variables — Method

dataset_variables(db::SQLite.DB, dataset)::AbstractDataFrame

Return the list of variables in a dataset

RDAIngest.death_in_ingest — Method

death_in_ingest(db, ingestion_id)

If ingested deaths are part of ingestion_id

RDAIngest.get_datasetname — Method

get_datasetname(db::ODBC.Connection, dataset)

Get the name of a dataset

RDAIngest.get_datasetname — Method

get_datasetname(db, dataset)

Return dataset name, given the dataset_id

RDAIngest.get_domain — Method

get_domain(db::DBInterface.Connection, domain_name::String)

Return the domainid for domain named `domainname`

RDAIngest.get_last_deathingest — Method

get_last_deathingest(db::ODBC.Connection, source_id::Integer)

Get ingestion id for latest death ingestion for source

RDAIngest.get_last_deathingest — Method

get_last_deathingest(source)

Get ingestion id for latest death ingestion for source

RDAIngest.get_namedkey — Method

get_namedkey(db::DBInterface.Connection, table, key, keycol)

Return the integer key from table table in column keycol (keycol must be a Symbol) for key with name key

RDAIngest.get_source — Method

get_source(db::DBInterface.Connection, name)

Return the source_id of source name, returns missing if source doesn't exist

RDAIngest.get_table — Method

get_table(db::ODBC.Connection, table::String)::AbstractDataFrame

Retrieve table table as a DataFrame from db

RDAIngest.get_table — Method

get_table(db::SQLite.DB, table::String)::AbstractDataFrame

Retrieve table table as a DataFrame from db

RDAIngest.get_valuetype — Method

get_valuetype(db::ODBC.Connection, variable_id::Integer)

Get the data type of a variable

RDAIngest.get_variable — Method

get_variable(db::DBInterface.Connection, domain, name)

Returns the variable_id of variable named name in domain with id domain

RDAIngest.get_variable — Method

get_variable(db::SQLite.DB, domain, name)

Returns the variable_id of variable named name in domain with id domain

RDAIngest.get_vocabulary — Method

get_vocabulary(variable, l)::Vocabulary

Get a vocabulary, name of vocabulary in line 1 of l, vocabulary items (code and description) in subsequent lines, comma-separated

RDAIngest.ingest_data — Method

ingest_data(ingest::Ingest, dbpath::String, dbname::String, datapath::String,
transformation_id::Integer, ingestion_id::Integer, death_ingestion_id=nothing)

Step 4: Import datasets, and link datasets to deaths

Ingestion_id can be from step 3 outputs if ingesting both death and datasets at the same time, if missing an ingestion will be created.

RDAIngest.ingest_deaths — Method

ingest_deaths(ingest::Ingest, dbpath::String, dbname::String, datapath::String; sqlite=true)

Step 3: Ingest deaths to deathrows, return transformationid and ingestionid

RDAIngest.ingest_dictionary — Method

ingest_dictionary(source::AbstractSource, dbpath::String, dbname::String, dictionarypath::String, datapath::String; sqlite=true)

Step 2: Ingest data dictionaries, add variables and vocabularies

RDAIngest.ingest_source — Method

ingest_source(source::AbstractSource, dbpath::String, dbname::String,
datapath::String; sqlite=true)

Step 1: Ingest macro data of sources: sites, instruments, protocols, ethics, vocabularies

datapath: root folder with data from all sources [DATAINGESTPATH]

RDAIngest.ingest_tac_vocabulary — Method

ingest_tac_vocabulary(source::AbstractSource, db, datapath)

TBW

RDAIngest.initstatuses — Method

initstatuses()

Default transformation statuses

RDAIngest.inittypes — Method

inittypes()

Default transformation types

RDAIngest.initvalue_types — Method

initvalue_types()

Add default value types

RDAIngest.insertdata — Method

insertdata(db::DBInterface.Connection, table, columns, values)

Insert a set of values into a table, columns list the names of the columns to insert, and values the values to insert

RDAIngest.insertwithidentity — Method

insertwithidentity(db::ODBC.Connection, table, columns, values, keycol)

Insert a record, returning the identity column value

RDAIngest.insertwithidentity — Method

insertwithidentity(db::SQLite.DB, table, columns, values, keycol)

Insert a record, returning the identity column value

RDAIngest.lines — Method

lines(str)

Returns an array of lines in str

RDAIngest.link_deathrows — Method

link_deathrows(db::ODBC.Connection, ingestion_id, dataset_id, death_identifier)

Insert records into deathrows table to link dataset dataset_id to deaths table. Limited to a specific ingest. death_identifier is the variable in the dataset that corresponds to the external_id of the death.

RDAIngest.link_deathrows — Method

link_deathrows(db::SQLite.DB, ingestion_id, dataset_id, death_identifier)

Insert records into deathrows table to link dataset dataset_id to deaths table. Limited to a specific ingest. death_identifier is the variable in the dataset that corresponds to the external_id of the death.

RDAIngest.link_instruments — Method

link_instruments(db::SQLite.DB, instrument_name, dataset_name)

Insert records into instrument_datasets table, linking datasets with instruments.

RDAIngest.lookup_variables — Method

lookup_variables(db, variable_names, domain)

Returns a DataFrame with dataset variable names and ids

RDAIngest.makeodbcparam — Method

makeodbcparam(s)

ODBC parameters are ? only instead of @name

RDAIngest.makeparam — Method

makeparam(s)

Prepend an @ to the column name to make it a parameter

RDAIngest.opendatabase — Method

opendatabase(path::String, name::String; sqlite = true)::DBInterface.Connection

If sqlite = true (default) open file on path as an SQLite database (assume .sqlite extension) else open database 'name' on server 'path' (assume SQL Server database)

RDAIngest.opensqlitedatabase — Method

opensqlitedatabase(path::String, name::String)::DBInterface.Connection

Open file on path as an SQLite database (assume .sqlite extension)

RDAIngest.opensqlserverdatabase — Method

opensqlserverdatabase(server::String, name::String)::DBInterface.Connection

Open database 'name' on server 'server' (assume SQL Server database)

RDAIngest.prepareinsertstatement — Method

prepareinsertstatement(db::ODBC.Connection, table, columns)

Prepare an insert statement for SQL Server into table for columns

RDAIngest.prepareinsertstatement — Method

prepareinsertstatement(db::SQLite.DB, table, columns)

Prepare an insert statement for SQLite into table for columns

RDAIngest.prepareselectstatement — Method

prepareselectstatement(db::SQLite.DB, table, columns::Vector{String}, filter::Vector{String})

Return a statement to select columns from a table, with 0 to n columns to filter on

RDAIngest.prepareselectstatement — Method

prepareselectstatement(db::SQLite.DB, table, columns::Vector{String}, filter::Vector{String})

Return a statement to select columns from a table, with 0 to n columns to filter on

RDAIngest.read_data — Method

read_data(datadoc)

Read file names and formatting parameters, returns a DataFrame with the data

RDAIngest.read_sitedata — Method

read_sitedata(source::AbstractSource, datapath, sourceid)

Data are aggregated by site_col in order to identify unique sites

RDAIngest.read_variables — Method

read_variables(source::AbstractSource, dictionarypath::String, dictionaryname::String)

Read a csv file listing variables, variable descriptions and data types in a dataset.

RDAIngest.save_dataset — Method

save_dataset(db::DBInterface.Connection, dataset::AbstractDataFrame, name::String, description::String,
domain_id::Integer, transformation_id::Integer, ingestion_id::Integer)::Integer

Insert dataframe containing dataset into RDA database and returns the dataset_id

RDAIngest.savedataframe — Method

savedataframe(con::DBInterface.Connection, df::AbstractDataFrame, table)

Save a DataFrame to a database table, the names of the dataframe columns should be identical to the table column names in the database

RDAIngest.savedataframe — Method

savedataframe(con::SQLite.DB, df::AbstractDataFrame, table)

Save a DataFrame to a database table, the names of the dataframe columns should be identical to the table column names in the database

RDAIngest.selectdataframe — Method

selectdataframe(db::ODBC.Connection, table::String, columns::Vector{String}, filter::Vector{String}, filtervalues::DBInterface.StatementParams)::AbstractDataFrame

Return a dataframe from a table, with 0 to n columns to filter on

RDAIngest.selectdataframe — Method

selectdataframe(db::SQLite.DB, table::String, columns::Vector{String}, filter::Vector{String}, filtervalues::DBInterface.StatementParams)::AbstractDataFrame

Return a dataframe from a table, with 0 to n columns to filter on

RDAIngest.selectsourcesites — Method

selectsourcesites(db::ODBC.Connection, source::AbstractSource)

Returns a dataframe with the sites associated with a source

RDAIngest.selectsourcesites — Method

selectsourcesites(db, source::AbstractSource)

Returns a dataframe with the sites associated with a source

RDAIngest.updatevariable_vocabulary — Method

updatevariable_vocabulary(db::DBInterface.Connection, name, domain_id, vocabulary_id)

Update variable vocabulary